Simplified code.

+++ Mcm.hpp
@@ -394,8 +394,8 @@ namespace WdRiscv
 
     /// Clear in the given mask, bits corresponding to the target
     /// instruction bytes covered by the given store instruction.
-    void clearMaskBitsForWrite(const McmInstr& storeInstr,
-			       const McmInstr& target, uint64_t& mask) const;
+    void clearMaskBitsForWrite(const McmInstr& storeInstr, const McmInstr& target,
+			       unsigned& mask) const;
 
     void cancelNonRetired(unsigned hartIx, uint64_t instrTag);
